Add login_commands to the auto_login schema
Accept a new parameter login_command in the lava_scheduler_app schema
which should contain commands to be run after logging in and before
starting any tests.  This is in particular useful to run commands such
as "sudo".

Also add a test case in lava_scheduler_app.tests.test_pipeline.

Change-Id: I2a8f0d702c7fbeb537cda23de46d35b58416469d
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>