Expand LXC support to add devices from all sybsystems
Allow all elements of a udev device to be added to the LXC by
not restricting the device list to the usb subsystem. Allows
USB serial devices to have the tty created in the LXC for support
of the ARM energy probe.
Rename the function to indicate it is not just the usb subsystem.

Change-Id: Iaee2424831a5f747c5685a8e6aab38f151469d54