api: add get_recent_jobs_for_device
This API function returns the last jobs asked for a given device as a list of
dictionaries containing the job ID, its description (the name), and its status.
The typical use case is to build a custom reporting tool outside LAVA by
querying this function, and then asking for job details once you know which job
to look for.

Change-Id: I2acd5415ed6ff90d2f64e40d520d79d09c2f4cf3
Signed-off-by: Florent Jacquet <florent.jacquet@free-electrons.com>