Add more Tegra124 based devices
All devices based on the Tegra124 chip have a common set of load devices
and at least all device we have in the Collabora lab use the same console by
default.

Add a common file, adjust the nyan file to be based on that and add
trivial configuration for the jetson-tk1.

Change-Id: Icc79ac3f433145b75daed35277bb638245a4c4ad
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>