Add collection of Exynos 4 and 5 devices
Most exynos 4 and 5 devices are similar with respect to
configuration.

None of the SoCs have direct ethernet support so usb is
needed for networking.

They do all tend to have OTG peripherals which can cause problems with
the default base ip args if ethernet gadget support is build into the
kernel (like it is in the exynos defconfig).

The load address are similar for all devices apart from exynos 5420
device for which the base ram address starts 512MB lower (allow for more
addressable physical memory without using lpae?).

This adds a common file for exynos 4 and 5 based SoC, with an exception
for 5420 (and it's 5800 chromebook specific derivative).

Apart from that the following device types are added/updated:
* Odroid X2
* Odroid XU3
* Snow (First Samsung ARM chromebook)
* Peach-pi (Update to use the common files)

Change-Id: Ia299ca4f403105a38fbcf5c4295ea55aaf924a88
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>