device-types: base-uboot: use run bootcmd
The u-boot command "boot" is an optional feature which is simply an
alias for "run bootcmd".  Several vendor uboot (Amlogic, Renesas) do
not have this command enabled, so using this command doesn't work.

To fix, just use "run bootcmd" directly, so the base-uboot device-type
can work on boards without this feature.

Also update the test case which checks for 'boot' in default args.

Change-Id: Iae28e22c0894d2567d93cbd5ee80b697883e40b0
Signed-off-by: Kevin Hilman <khilman@baylibre.com>