Fix validation of boot auto_login parameters and add a unit test
The AutoLoginAction.validate method checks for the presence of the
login_prompt parameter then checks it is not empty.  An exception is
raised when the parameter is not present:

  KeyError: 'login_prompt'

Fix this by checking that login_prompt is not empty only if it is
present in the dictionary.

Also add a new test module to cover the auto_login parameters of the
pipeline boot action, and add it to the list of tests.

Change-Id: I922bfdb03734f99a872ab4ea4028339135bb19ef
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>