device-types: add sun6i-a31-app4-evb1
This adds device type for the Allwinner A31 APP4 EVB1 Evaluation Board
(sun6i-a31-app4-evb1) which is U-Boot based.

The Ethernet is provided by a USB to Ethernet adapter with a modified
version of U-Boot. It has to detect USB devices before trying DHCP and
assign an Ethernet interface.
To have Ethernet over USB working in U-Boot, you have to add the
following lines to one of the header used to create the U-Boot version
specific to your board (can be sunxi-common.h for example):
 #define CONFIG_USB_HOST_ETHER
 #define CONFIG_USB_ETHER_ASIX
Check NET, PING, DCHP and USB_EHCI_HCD in menuconfig.

The Ethernet interface is automatically set to asx0 (the USB to Ethernet
adapter).

Change-Id: I1929f57a4d75b2bb12eb58e12e1b29a7355e0c8b
Signed-off-by: Florent Jacquet <florent.jacquet@free-electrons.com>