Fix typo which causes XML-RPC error when saving device dict
Fix "raise xmlrpclib.Fault(..." in lava_scheduler_app/api.py.  The
error seen was:

  <Fault -32603: "Internal Server Error (contact server administrator for details): 'module' object has no attribute 'Faul'">

This code is only run when the device dictionary fails to be saved on
the disk, for example due to a permission error if /etc/ is owned by
root but the server is running as lavaserver.

Fixes: ae4d4776ca7b7454f5406159226e3c9327dd207f
Change-Id: Ic5c480bf610b9e87cdb74502fcd6972c496400aa
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>