device-types: add bcm2837-rpi-3-b
Add device-type to support the Raspberry Pi 3b running mainline u-boot.

Note that this intentionally does not use the rpi-common.jinja2 because
rpi3 is 64-bit an therefore uses 'booti', and having bootm_* bootz_* and
booti_* all in the device-type doesn't work and LAVA always uses bootm,
and also the rpi3 uses ttyS0 as a console and rpi-common sets ttyAMA0.

Also, the addresses used in rpi-commmon don't allow enough room for
large kernel images (uncompressed in arm64), so the ramdisk addr is
moved a bit higher than in rpi-common.

Note that boot_character_delay is needed for this board because by
default, with mainline u-boot, the stdout environment variable is set to
"serial,lcd" which means u-boot ouput is displayed on both the
display (HDMI) and the serial port.  With that set, the uboot UART
driver drops characters and causes lots of problems for LAVA.

An alternate approach is to disable uboot output to HDMI using
"setenv stdout serial; env save".  If that is done, the
boot_character_delay could be removed for boards where that is done,
but let's keep the default device-type safe for all u-boot settings.

Tested on an rpi 3b running mainline u-boot (64-bit) from v2017.05
(without changing the stdout env variable.)

Change-Id: I017a0c051ecba68eee0e1d5d0e8f0b6a5eb51a35
Signed-off-by: Kevin Hilman <khilman@baylibre.com>