Add flag to disable uniquifying the download paths
Add support for keeping artifacts in the indicated path rather than
further uniquifying the path by appending the key, using an extra
argument `uniquify` in DownloaderAction.  This is useful for actions
that want to download a set of artifacts into a single directory and
make them available to the device under test.

Change-Id: Ieae158e87c4cc23631506d198dc680875a36a4bc
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>