Allow configurable writes for secondary storage deployment
dd is very common tool to write images to external storage, but it's
not a great choice in general.  For SD/USB media something like
bmaptool is a much better choice, for e.g. raw NAND specialized tools
are needed.

Address this by making the writer tool configurable as well so the
test writer can choose the best tool for the situation.

The way it's set up is to add a writer parameter block similar to the
current downloader one.  If only the downloader is configured the
behaviour is the same as currently, if only the writer is configured
it's assumed the writer tool also does the downloading, if both are
configured they're run in a shell pipeline: "downloader | writer".

Furthermore support is added to provide a list of images for the
dispatcher to download and expose to the board in a single directory
over HTTP.  They may include an overlay for the target image via the
apply-overlay key.  The DOWNLOAD_URL substitution will be the URL of
the image with the "image" key.  In the bmaptool case the image can be
the actual image, with the meta-data URL file being automatically
determined from it by the tool.  For other tools, it could be a
manifest file.

Note: unit tests and documentation may need to be updated

Change-Id: I4ac1cf7289f3b2372fb35c5b9eb609cf03482edf
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>