Add option to enable listening on IPv6
Add a command line option --ipv6 to the dispatcher-master command to
enable IPv6 on the ZMQ listening sockets for both the master and
logger.  When this is enabled, they will be listening for both IPv4 and
IPv6 incoming connections.

Also update the init script and default config file to add the option
with some examples.

Tested successfully on Debian Stretch with a standalone set-up.

Change-Id: I0a93492b9ca25c625f65b2a383347a6891249a73
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>