Rework test shell feedback timeouts
The goal of this patch is to:
1/ decrease the time between two iterations of the loop
2/ lower the timeout when listening to the feedbacks
3/ grab feedback at regular intervals

1/ We should try to decrease the time between two iterations of the loop to have
regular feedbacks. Currently we will wait up to 150 seconds if no patterns
match.

2/ Their is no reason to listen for more than some seconds for the feedbacks.
The goal is mainly to purge the buffer of characters. These characters are
already present in the buffer and will be returned immediately.

3/ We should listen to the feedbacks only after test_connection.timeout seconds
and not for each iteration of the loop. In fact, _keep_running returns for
every match (results, ...).

Change-Id: I157032f299a2b52908107558d164da1873e6bcdb