ramdisk: only pass to bootx if has u-boot header
Whether or not the ramdisk has a u-boot header is an option in the job
parameters.  If add_header != u-boot, then the ramdisk should not be
passed to the u-boot bootm/bootz/booti commands.

This feature is useful because a ramdisk image (without u-boot header)
can be passed to the kernel using DT properties instead of via u-boot.

Change-Id: I84f228118c8713522856c48d85c16c09daab6b7c
Signed-off-by: Kevin Hilman <khilman@baylibre.com>