dispatcher-master: prevent growing message queue
This patch makes the main loop to process all the pending messages in the
controller socket before processing the jobs which can lead to generate many
messages. Thus, we ensure that the controller message queue won't grow too large
if the processing of the jobs begins to take a lot of time for any reason.

Hacked-by: Rémi Duraffort <remi.duraffort@linaro.org>

Change-Id: I9e4fdd094276837f1fa63506c7b9648f99d46a8c