results_app: API: add the 'make_custom_query' function
This adds a 'make_custom_query' function to the lava_results_app API, to allow
queries to be performed by scripts.

This basically takes the queried entity as first argument, and the conditions as
second one. A list of dictionaries with the results is returned. The
dictionaries depend on the entity queried (see the Query documentation for more
informations).

Change-Id: I5a9fdf555b281dd116bc7f5dada39bccd8122ff1
Signed-off-by: Florent Jacquet <florent.jacquet@free-electrons.com>