results_app: API: add the 'run_query' function
This adds a 'run_query' function to the lava_results_app API, to allow scripts
to get the result of a query.

This basically takes the query name as first argument. A limit and a username
can also be specified to restrict the number of results, or if the query owner
is not the same as the authenticated user querying the results.

Change-Id: Ib8bb1e96c012682da934c1657860c271f2cc01b5
Signed-off-by: Florent Jacquet <florent.jacquet@free-electrons.com>