XMLRPC: add offline_info argument to list endpoint
Add an offline_info argument to list endpoint to get more infos on the
returned devices that are offline. If offline_info is True (defaults to
False) and the device being iterated over is offline, the date since
which it is offline and the user responsible of this offlining are added
to the dictionary of the device.

Change-Id: Idc66e7c11c089a9611b18c3245f1e3dad42e1f67
Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>
(cherry picked from commit 25eaab17618e672aead2a360743f255b31ec45f0)