Stop parsing kernel messages when the end of a panic has been found
When parsing the kernel log in auto_login, kernel panics get matched
correctly and the errors are added to the results.  Then as panics are
fatal, it needs to stop waiting for some more messages and return.  If
it keeps waiting then the job will time out and the parsed errors will
be lost.

To solve this issue, abort parsing the kernel messages as soon as the
end of a kernel panic has been matched.

Change-Id: If574b2c287dcade2dcde5361d0c399338df94591
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>