Document option for device_type qemu to choose host audio handling
If using the device_type qemu and passing -soundhw XYZ (e.g. -soundhw hda) in extra_options:
then qemu will look for sound hardware (usually pulseaudio) on the host running qemu.
This is:
- either not available on a server
- or not intended to get out of the VM on a worker node

Thus document the environment variable in env.yaml to change the qemu
audio driver. For more options see
  qemu-system-x86_64 --audio-help

v2: Add it to the documentation pages and also a note in the jinja2 template
v3: Include feedback

Change-Id: I52b56850176426e180e5f49b53ac9030311414fd
Signed-off-by: Jan-Simon Möller <dl9pf@gmx.de>