Add option to enable IPv6 connections
Add a command line option --ipv6 to the lava-dispatcher-slave command
to enable IPv6 on the ZMQ connection sockets.  This applies to
connections to the LAVA master as well as the logger.

Also update the init script and default config file to add the option
with some examples.

This relies on a corresponding change in lava-server to enable
listening on IPv6.

Tested successfully on Debian Stretch with a stand-alone set-up.

Change-Id: I1cc27624c804ba8f684c0901724c26b1b5ee2513
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>