Provide useful output when no classes accept an action
If any of the actions in a job are not accepted by any pipeline
class, log the reasons for each class to allow job writers or
developers to better understand why it has been rejected.
If a deploy/boot/test class does not return a tuple for it's
accept function, throw a LAVABug exception.

Change-Id: I0365c12f64b13a07217794e52908ceba55341206