Fix SQL request storm when listing jobs
Using select_related and prefetch_related remove some requests. But the only
way to dramatically reduce the number of request is to use Django cache.

We now cache in memory (for 30 seconds) the result of
DeviceType.some_devices_visible_to which is called a lot for the same device
type.

With this patch, we go from 206 to 6 queries when rendering /scheduler/alljobs.

Change-Id: Ib1e3df695a0acd4e450faaa8699baa88380bcc3a