LAVA-1053 Results limit does not work for queries
While this cannot be fixed for live queries, it works for
materialized view queries.
Also removed some of the unnecessary code.
The reason it cannot work for live is that django cannot support
modifying query sets after applying limit to the query - which
is what we do in lava tables.
Updated api calls to use limit as well.

Change-Id: I3d1dad5fd7a38d28155334181ebfa34952262bb2