Do not rely on SysV runlevel to ensure container is ready.
In some operating systems (archlinux) which uses systemd the SysV runlevel
is not being mapped properly and always gives an 'unknown' result. Hence do
not rely on it, instead use lxc-info to find out the state of the container
and also based on IP address obtained by the container we can ensure that
networking is enabled and up.

Change-Id: I6ee35ebfe9a270d57200e1eb697fcd77af83922e