Add option to not uniquify secondary image download paths
Some tools such as bmaptool need to download several image files from
a common URL path.  Having each file stored in a separate directory on
the dispatcher breaks this concept and there is no easy alternative to
specify different URLs for each file to be downloaded onto the board.

Remove the hard-coded 'image' suffix in the image download path on the
dispatcher to avoid this issue and add a `uniquify' parameter to the
deploy action.  This can be set to false in which case all the image
files are available in the same path on the dispatcher.  It makes use
of the new uniquify=True option with DownloaderAction for secondary
deployments.

Change-Id: Ia3c335e29805937c15da51498678c0b065306cc1
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>