Revert "Stop parsing kernel messages when the end of a panic has been found"
This reverts commit ce2dbd83fef8e4fa47261eb41bf580d1ed29eeac.

While this change makes sense on its own, it then exposes the issue
whereby the job stops but does not fail to log in, so it aborts
without setting the status to "Imcomplete".  Another fix should be
made to abort the job when a panic has been met.  In the meantime,
revert to the old behaviour of timing out on the login prompt when the
kernel fails to boot.

Change-Id: I962947604a2f588bd4352b19403497a844313eef
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>
(cherry picked from commit d91a5b9984ebeaa8ca39c3326531072c3e020831)