Move lxc-static action to a boot method
Devices can only be added to the LXC after running lxc-start,
so move the static devices action to a boot method, after
LxcStartAction. Fix handling of device_info and static_info
so that if a device_info is passed to the function, that overrides
the one from the job which is the dynamic info.

Change-Id: I3649df7439de7a65cbe45e70748cc6ddf1038e9c