Fix a data race between lava-logs and lava-master
If a test job is finishing really quickly, two messages will be send at the same time:
1/ START_OK lava-slave => lava-master: mark the job as RUNNING
2/ lava.job result lava-run => lava-logs: mark the job as finished (COMPLETE or
   INCOMPLETE)

The two messages are handled by two different processes, so the order cannot be guaranteed.
If the first message is handled after the second one, the job will be marked as
RUNNING while it has been finished.

Change-Id: Ibb3f5fc26eb53527c98a0954439ccc231062a49e