Add base-depthcharge and rk3288-veyron-jaq device types
Add base-depthcharge.jinja2 as a first device type definition to use
Coreboot with Depthcharge (see matching dispatcher changes to add the
"depthcharge" boot method).  This includes commands for nfs and
ramdisk.  The kernel image needs to be an FIT image.

It relies on Depthcharge to be configured with its command line
interface enabled and the "tftpboot" command available.

Add the rk3288-veyron-jaq device type definition based on it.

Add a template test for base-depthcharge.jinja2.

Change-Id: I82183e10ea0b86baa68c04831bd7f7babfee6c13
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>