Generate the FIT image for Depthcharge
Add a PrepareFITAction class to create an FIT image using mkimage.
Unlike uImage, the FIT image may also contain a dtb and a ramdisk.  So
it needs to be run after the ramdisk has been recreated with an
overlay.  The downloaded kernel image now needs to be a plain image;
this removes the option for booting FIT images directly.

Add a PrepareFITAction pipeline stage via PrepareKernelAction at the
end of the PrepareOverlayTftp action if the boot method is
`depthcharge'.

Also check that `mkimage' is available on the system, raise an
infrastructure error exception otherwise.  The `mkimage -f auto'
option is supported in mkimage 2016.11 (Debian Stretch) but not in
version 2014.10 (Debian Jessie).

Change-Id: I55c1d309d330eafb197b2642dcf477a44fedc52f
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>