Add FIT parameters to base-depthcharge and rk3288-veyron-jaq
In order to generate an FIT image, the CPU architecture and kernel
load address need to be defined in the device parameters.  Add these
in base-depthcharge and rk3288-veyron-jaq.

The tftpboot command line does not take the path to the plain kernel
image any more but to the generated FIT image instead, so update the
template accordingly.

Update the template test accordingly as well.

Change-Id: I988ac9ef168f7029dee26ba3ecde0e71ebb87eec
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>