Disable high limits on ODROID-XU3
With kernel version v4.8 and earlier (including stable branches), a
memory region is reserved for the MFC hardware codec at a fixed
address.  This conflicts with the loading address of the ramdisk if it
is not relocated and causes a failure to boot.  While it's possible to
change the device tree to make the reserved memory regions dynamic
like in more recent kernels, it's overall more appropriate to fix the
device configuration to relocate the ramdisk and dtb.

Disable base_high_limits on odroid-xu3 to fix this issue.

Change-Id: I1229b44a7bca658f1fb328464d7ee7472c5f74a6
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>
(cherry picked from commit ff1ef1c4dcbd2c9771d5e61f55b17f666b3f057d)