device-types: add rk3399-gru-kevin
This new device type is for a gru-kevin Chromebook with mainline
Coreboot/Depthcharge bootloader built with the debug CLI enabled to
use the tftpboot command.  It needs a Servo Micro debug board and a
USB Ethernet adapter, either using a separate Servo v2 board or any
Linksys dongle based on AX88772.

Change-Id: I5c64ac1f6e131f9082478ea880ae088cc19926bc
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>