device-types: fix broken conditional in base-depthcharge
The fit_kernel_load_address parameter is required for Depthcharge and
has a default value of 0 so it should not be within a conditional
block.  In fact this breaks all the Depthcharge device configs that
use the default load address (i.e. all the current ones).

Fix this by removing the conditional, which should not cause any unit
test failures.

Change-Id: I478cdbeffee4419f3bc95af33a622c87f775307e
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>
Fixes: f61721206873 ("Ensure added blocks are wrapped in conditionals")