Fix test for extra results metadata in callback data
The callback results data includes some metadata which in turn
includes some arbitrary and optional extra data parsed from a YAML
file generated by the test job.  The string with the path to this
optional YAML can be either a `str' or a `unicode' object, and due to
other changes was a `str' in 2017.12 but is now a `unicode' in 2018.2.
So the extra data is not loaded from the YAML file any more; instead
the path to the file is left in the callback data.

Fix this by testing if the extra data path is a `basestring' instance
which is a parent type for both `str' and `unicode'.  Also remove the
redundant check if extra_data is empty since None is not a basestring,
and a file path with an empty string does not exist.

Change-Id: Ia06929aa214584f515d11870fa391cf4db8e770f
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>