Pass LXC/fastboot parameters to all jobs in split_multinode_yaml
reboot_to_fastboot was not handled before, so that Android devices would
always reboot to fastboot after test runs, even if this parameter was
set to false in the job submission.

lava-lxc protocol metadata was only populated for the first job of each
role. This change passes the data to all job instances, so that multiple
jobs per role can be created.

Change-Id: I5a89d748468e8b85224c2c60be8c8016b203b648