Misc beaglebone black fixes
My BBB was failing to boot a uImage due to an issue with the addresses
(I didn't dive into the details why). Changing the bootm_ addresses to
match the bootz_ addresses (as has been done on various other boards)
solved the issue.

Further more in some cases during auto-login the username was received
as 'suer' rather then 'user', add a small boot character delay to
resolve that.

Change-Id: I84431b7d037623c39c8d25dce1a0d603a3867cd0
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>