Allow dispatcher unit tests to generate device config
With the combined codebase, it is now possible to generate the
device configuration for each lava_dispatcher unit test directly
from the current templates in lava_scheduler_app so that changes
in the templates are picked up in the dispatcher tests.
This will allow replacement of static YAML in the devices directory
with device dictionary Jinja2.
Also remove a reference to example.com which can fail the MPS2 test.

Change-Id: I7c63acf857b5d22bb0f0a519ba955e1de282a536