LAVA-1173 - allow for matching parents of nodes
To match devices other than /dev/bus/usb/, the match needs to
be made on the parent device (determined in udev by matching
ATTRS{} instead of ATTR{}). This allows /dev/ttyUSB4 to be
matched as a child of a device with matching ATTRS{} when the
tty device itself does not have matching ATTR{} values.
The device_info needs to contain a key:value pair of
'parent': True as well as the vendor and product id pairs.

Change-Id: Ic36d895f84ade1a4ca731a31af9b87a61528e259