LAVA-1316 Instant scheduling
When a device becomes available or a job is submitted, a zmq signal is sent.
lava-master is listening for such events and can starts the scheduler whenever needed.

lava-master will wait for one second before starting the scheduler to leave
some time to connect all events.

If a user is submitting a lot of jobs, this will allow to call the scheduler only once.

Change-Id: Ia49d9612177da968a2895175760445d668ed8e1a