Use /dev/kmsg for LAVA test signals, if requested
Adds a parameter "lava-signal" to the test definition parameters
which when set to "kmsg" will tell the lava test scripts to echo
their output to /dev/kmsg.
This *might* help with test cases that produce kernel output and
can interrupt LAVA signals sent to stdout.

This review does NOT cover the issue that lava-test-runner signals
will still only be sent to stdout.

Change-Id: I8ac92c725ac49e40847d1a54c13bef81030903fa