Fix map_scanned_results() when test status string is invalid
When a test result produces an invalid status string, the correct
value to return is None.  By returning False, lava-logs does not treat
it as invalid and adds a False object in its list of test cases which
then results in an uncaught exception when calling bulk_create():

  File "/usr/lib/python3/dist-packages/django/db/models/query.py", line 404, in _populate_pk_values
      if obj.pk is None:
      AttributeError: 'bool' object has no attribute 'pk'

Fix this by returning None rather than False when the status string is
invalid.

Change-Id: I080d1748998926725e3ea5f73487da1c32449ca6
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>