Add gitlab ci configuration
Expand the start of the current Developer Guide to include
the Contribution Guide.
Change advice on branches and workflow to the Community
Project workflow in gitlab.
Update references to files in the git repo.
Update README for GitLab
Update the support page for lavasoftware.org
Update README for GitLab
Make it clearer that the CoC is not just for communications on mailing
lists etc., but for all behaviour around the LSCP.

Signed-off-by: Neil Williams <neil.williams@linaro.org>