Check tag support for multiple devices per role
The tag check when checking availability prior to creating the test job
only checks if there is at least one device that satisfies the tag
requirements. However, it is possible to need more than one device with
the same tags in the case of a multinode job where the count is > 1 for
any of the roles. The fix is to pass in the required number of devices
that satisfy all the tags.

Change-Id: I8c2f88703c33256bb61f64179a9905126cc0e649