Remove unnecessary Python 2 imports
In Python 3, all the strings are of the `str' type.  Python 2.x used
to have `basestring' but as this code base now only supports Python 3,
there is no need to be compatible with Python 2.x.  Remove the
`basestring' reference to `str' to reflect this.

Also remove the Python 2.x specific import of `quote' for similar
reasons.

Change-Id: I8b6f329e40bcd57ed82360caed8a25a2af190909
Fixes: 0d1e0a0705e0 ("Fix test for extra results metadata in callback data")
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>