scheduler: randomise the device list
Currently, with N devices of the same device-type, all public and all
available, when scheduling num(jobs) < N, the exact same set of devices will be
always used while the remaining will be idle.

By making the sort partially random, we split the load accross all the devices,
even when num(jobs) < N.
BUT, this is making the scheduler a bit less reproductible.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>
Tested-by: _Marge_Bot <https://git.lavasoftware.org/ivoire/lava/merge_requests/2>