Support for skipping test shell timeout
Add support for tags in the test: action with the first tag
timeout with 'skip' set to allow the test job to continue to
the next boot: action block if a test: action block times out.
Not supportable for MultiNode. Closes: #81
Some pylint3 fixes.
Ensure the new unit test is validated against the submission
schema.

Signed-off-by: Neil Williams <neil.williams@linaro.org>