Make primary key non-editable
This is confusing most admins because this is not doing what they expect.
When modifying the primary_key of an object, Django will make a copy of the
object and change the primary_key of the copy, keeping the original.

For instance, when a TestJob is running on a Device, changing the Device
hostname will create a copy of the Running Device without any TestJob attached
to it.

Change-Id: Ic7329156c5f45ccdeb42c1a128981d3e36a9d64d