Create but not edit objects in admin interface
Device, DeviceType and Worker need to be created in the Django
Admin Interface but not allow editing after creation.
Change the editable field back but change admin.py to support
changing the read_only status based on the call being create
or edit.
Move latest migration to replace non-editable key changes
which are no longer required.

Change-Id: Ie5f565d300408e4697738322ec1e42ec42b50ec9