Allow a user-specified kernel failure message
Support for devices or kernels which do not reliably output
predictable messages when the boot fails and the device
resets.
Test jobs can specify a failure_message which, if matched
in the kernel boot messages, will fail the job.
Ensure fail_msg only triggers when set
Ensure the job ends when the failure message matches.
Check for the failure_message before success
Add documentation of the new string. Closes: #75

Signed-off-by: Neil Williams <neil.williams@linaro.org>