Implement split behaviour for Debian backports
Avoid duplicating listings by conditionally repeating the
call for the dependencies.
If used for package building, create a combined list
covering backports and the corresponding stable suite.
If used with --names, only output the dependencies which
need to be installed from backports and require a second
call by the user to get the list of dependencies needed
from the corresponding stable branch. (The call to install
from backports requires different command line options.)
Refactor to put this specialist knowledge into a specific
function for Debian, allow for other functions for other
distributions.

Signed-off-by: Neil Williams <neil.williams@linaro.org>