Explicitly set autoescape to False for YAML.
Jinja2 autoescape defaults to False but this default is
likely to change. Autoescape behaviour is predicated on
HTML for XSS vulnerabilities, so would be set by Django
in handling the UI templates. HTML (escaped or not) would
be either invalid YAML or ignored comments.
Autoescape also has a performance penalty when enabled,
so set to False now to protect against the future
change in the default in jinja2.

Signed-off-by: Neil Williams <neil.williams@linaro.org>