Add output about the build to the logs
Each run only builds once, so the first .changes file
in /tmp is the right one. Find it and pass it to
debc which shows the content of all packages to provide
confirmation that any package dependency changes have
taken effect. debc is part of devscripts which already
exists to run debian-dev-build.sh

Signed-off-by: Neil Williams <neil.williams@linaro.org>