download: Raise an Infrastructure Error on timeout
In fact, the validation() function does already check that the resource exists.
So if downloading fails, it's likelly to be an issue with the infrastructure.

Also fix the timeout duration when retrying.
The fix is in fact not perfect because we have no way to get the parents
timeouts that might be smaller.

Closes #131