lava_results_app: add test case for YAML Decimal object conversion
The test case metadata may contain measurements which are stored as
Decimal objects in the database.  By default, they get dumped in YAML
as a Decimal Python object which makes it incompatible with standard
parsers when the results are exported (in YAML or CSV).

Add a test case to catch this problem with arbitrary metadata,
checking the actual YAML text stored in the TestCase object.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>