Extend 0f7a7ec8 to other command_output comparisons
Not all commands provide sane exit codes but when the output of
the command must be parsed, ensure that the string match is
short and case-insensitive to protect against some of the
likely optimisations in the tool used in the command.
Add a new parsed_command function to Action to help
future debugging.

Signed-off-by: Neil Williams <neil.williams@linaro.org>