interactive: allow to only wait, not sending any command
For some boards (like frdm), the boot does not wait for any prompt. So the
interactive script should start by waiting for a prompt (or a message).
Also mention in the documentation that by default the script will not wait for
a prompt for the last command (if a message is matched).

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>
Reviewed-by: Neil Williams <neil.williams@linaro.org>
Tested-by: _Marge_Bot <https://git.lavasoftware.org/lava/lava/merge_requests/255>