Add a function to wait for the django migrations
By default, the entrypoint will wait for postgresql then run the migrations.
If LAVA_DB_MIGRATE is set to "no", the entrypoint will wait for postgresql and
then wait for the migrations to be executed by another container.

This allow to have one service per container and only one service running the migrations.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>