Add --host argument to pg_isready command call
The current call expect a local database. It shouldn't be assumed local as
it can be a remote database. Pass --host argument, similar to the psycopg2
connection. In the case the database is local, we'll use 'localhost' and
will work the same way as before.

In addition, use the long form of the command option, more readable.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>