doc: make the message toward 3rd party packaging more friendly
The messaging toward packaging and shipping LAVA differently than
the officially supported method (i.e. Debian) is threatening.
Keep it clear that there's an official method supported and other
methods are still welcome if someone is willing to do it.
It doesn't mean another method will be officially supported but
it doesn't shutdown a contributor to do as he wishes either.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>