device-types: base: detect TFTP ERROR: File not found
When runnning a uboot job, our lab hit this issue:
Filename '1`0/tftp-deploy-7ong68s3/kernel/uImage'.
Load address: 0x1080000
Loading: *T
TFTP error: 'File not found' (1)
Not retrying...

But this error was not catched by LAVA and so the job continued
and started with an empty kernel up to a "wait for prompt timed out"
error.

This patch adds a way to detect this infrastructure error.