device-types: minnowboard-common: permit to override grub_autoboot_prompt
The default grub_autoboot_prompt is bad, since it detect the bios
startup and not the grub display.
This lead to job failure on my minnowboard when using grub over TFTP,
since the interrupt sent by LAVA came too early.

But by fear to break existing users, the simplier way is to add a way to
override the default grub_autoboot_prompt.