Set lavaserver uid and gid to a known value
When upgrading from one container version to another, the uid/gid should not changed.
In fact, if uid/gid are changing, the files/directories lava service will fail
to update the files/directories created by a previous container version (job
logs, device dicts, ...)

We are choosing a different value than the current (uid,gid) but this will be
better if, in the future, more users are created.

Closes #12

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>