Revert "device-types: base: detect TFTP ERROR: File not found"
This reverts commit 668d132c67d34974c8b0e492203cb4e584139292.

With this patch applied, some jobs are going to fail (found by meta-lava).
Jobs like https://staging.validation.linaro.org/scheduler/job/249493#L271 (that
are not giving any RAMDISK) will fail while they should not.

This patch will be reintroduced for the next release along with some work to
fix the crashes that it does imply.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>