Add overriding resets_after_flash to support newer FRDM-K64F boards
Newer debug firmware on FRDM-K64F doesn't "reset after flash" when
programming the flash via USB Mass storage.  To support boards that have
the older and newer firmware we need to allow 'resets_after_flash' to be
set on a board by board basis.  We expose the 'resets_after_flash'
paramater so that on newer boards we can set this to 'false' and we
default to 'True' so older boards still function as they always have.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>