fix pyocd boot action method to print out version of binary used
This commit changes the boot action method to print out the version of the
pyocd-flashtool utility used by the method. It is non-Debian so it was
previously failing with the original methodology of querying the version
number using debian_filename_version(). It now uses a new function
added in utils.py to run the utility to extract the version info.

Signed-off-by: Vincent Wan <vincent.wan@linaro.org>