LAVA branding: add possibility to include custom CSS
This patch adds the possibility to include a custom style sheet in a
LAVA instance for applying an overall branding to the GUI.
The CSS file is configurable via the BRANDING_CSS variable in
/etc/lava-server/settings.conf. It is added at the very bottom of CSS
includes in the base template in order to make it override all
previously included style sheets.
Some styles are defined inline within the base template. In order to
override these, the CSS directive "!important" has to be added to them
in the custom style sheet.

Signed-off-by: Tim Jaacks <tim.jaacks@garz-fricke.com>